requirements.txt
# Core Testing Framework
pytest>=7.4.0
pytest-html>=3.2.0
pytest-json-report>=1.5.0

# Data Parsing
pyyaml>=6.0
jsonschema>=4.17.0

# Network Testing (optional - install as needed)
# pyats>=23.0
# genie>=23.0
# napalm>=4.0.0
# netmiko>=4.1.0
# textfsm>=1.1.0
# ttp>=0.9.0

# Traffic Analysis (optional - install as needed)
# pybatfish>=2023.0

# Linting and Code Quality (dev dependencies)
# black>=23.0.0
# flake8>=6.0.0
# pylint>=2.17.0
